refactor(lib): move util
