+++ app/resources/v1/street_remixes.js
@@ -1,7 +1,7 @@
 const { Street } = require('../../db/models')
-const { ERRORS } = require('../../../lib/util')
 const logger = require('../../lib/logger.js')
 const { streetsToCSV } = require('../../lib/streets_export.js')
+const { ERRORS } = require('../../lib/util')
 
 exports.get = async function (req, res) {
   // Flag error if user ID is not provided
